Api middleware handles the api calls.
It transparently gets passed a request without its uri prefix, so the run function thinks it is root.
This way we could just run the api if we wanted to without the UI.

I want to also use functors to be able to nest middlewares, however I don't know a way
to have this recursion with a base case with functors. For now this relationship is hard coded.